trie:index
SlimIndex provides a handy API for most common usages.
One of the most common case is to use SlimTrie to index on-disk data.

A SlimIndex instance contains a SlimTrie index and a user defined
DataReader which is used to load data from storage, such as from a file.

SlimIndex is introduced to make it easier to use for end user.