fix: scan trie with one key
Problem:

If there is only one key, the walking algo does not work,
because it stores scanning progress on parent node.
A one key trie does not have a parent node thus it just return nothing.

Solution:

Treat a one key trie a special case and return an iterator that just returns
the only one record.

- Fix: #170